+++ source/Public/New-SqlDscDatabase.ps1
@@ -33,6 +33,13 @@
     .PARAMETER OwnerName
         Specifies the name of the login that should be the owner of the database.
 
+    .PARAMETER IsLedger
+        Specifies whether to create a ledger database. Ledger databases provide
+        tamper-evidence capabilities and are immutable once created. This parameter
+        can only be set during database creation - ledger status cannot be changed
+        after the database is created. This parameter requires SQL Server 2022
+        (version 16) or later, or Azure SQL Database.
+
     .PARAMETER DatabaseSnapshotBaseName
         Specifies the name of the source database from which to create a snapshot.
         When this parameter is specified, a database snapshot will be created instead
@@ -137,6 +144,10 @@ function New-SqlDscDatabase
         [Parameter(ParameterSetName = 'Database')]
         [System.String]
         $OwnerName,
+
+        [Parameter(ParameterSetName = 'Database')]
+        [System.Boolean]
+        $IsLedger,
 
         [Parameter(Mandatory = $true, ParameterSetName = 'Snapshot')]
         [ValidateNotNullOrEmpty()]
@@ -255,6 +266,24 @@ function New-SqlDscDatabase
             }
         }
 
+        # Validate IsLedger if specified (requires SQL Server 2022+)
+        if ($PSBoundParameters.ContainsKey('IsLedger'))
+        {
+            if ($ServerObject.VersionMajor -lt 16)
+            {
+                $errorMessage = $script:localizedData.Database_IsLedgerNotSupported -f $ServerObject.InstanceName, $ServerObject.VersionMajor
+
+                $PSCmdlet.ThrowTerminatingError(
+                    [System.Management.Automation.ErrorRecord]::new(
+                        [System.InvalidOperationException]::new($errorMessage),
+                        'NSD0007', # cspell: disable-line
+                        [System.Management.Automation.ErrorCategory]::InvalidOperation,
+                        $IsLedger
+                    )
+                )
+            }
+        }
+
         # Validate source database exists when creating a snapshot
         if ($PSCmdlet.ParameterSetName -eq 'Snapshot')
         {
@@ -311,6 +340,11 @@ function New-SqlDscDatabase
                     if ($PSBoundParameters.ContainsKey('CompatibilityLevel'))
                     {
                         $sqlDatabaseObjectToCreate.CompatibilityLevel = $CompatibilityLevel
+                    }
+
+                    if ($PSBoundParameters.ContainsKey('IsLedger'))
+                    {
+                        $sqlDatabaseObjectToCreate.IsLedger = $IsLedger
                     }
                 }