+++ tests/Unit/Public/New-SqlDscDatabase.Tests.ps1
@@ -73,6 +73,7 @@ Describe 'New-SqlDscDatabase' -Tag 'Public' {
                 $mockDatabaseObject | Add-Member -MemberType 'NoteProperty' -Name 'RecoveryModel' -Value $null -Force
                 $mockDatabaseObject | Add-Member -MemberType 'NoteProperty' -Name 'Collation' -Value $null -Force
                 $mockDatabaseObject | Add-Member -MemberType 'NoteProperty' -Name 'CompatibilityLevel' -Value $null -Force
+                $mockDatabaseObject | Add-Member -MemberType 'NoteProperty' -Name 'IsLedger' -Value $false -Force
                 $mockDatabaseObject | Add-Member -MemberType 'ScriptMethod' -Name 'Create' -Value {
                     # Mock implementation
                 } -Force
@@ -100,6 +101,24 @@ Describe 'New-SqlDscDatabase' -Tag 'Public' {
             $result.RecoveryModel | Should -Be 'Simple'
             $result.Collation | Should -Be 'SQL_Latin1_General_CP1_CI_AS'
             $result.CompatibilityLevel | Should -Be 'Version150'
+        }
+
+        It 'Should create a ledger database with IsLedger set to true' {
+            # Create a mock server for SQL Server 2022 (version 16) which supports IsLedger
+            $mockServerObject2022 = New-Object -TypeName 'Microsoft.SqlServer.Management.Smo.Server'
+            $mockServerObject2022 | Add-Member -MemberType 'NoteProperty' -Name 'InstanceName' -Value 'TestInstance2022' -Force
+            $mockServerObject2022 | Add-Member -MemberType 'NoteProperty' -Name 'VersionMajor' -Value 16 -Force
+            $mockServerObject2022 | Add-Member -MemberType 'ScriptProperty' -Name 'Databases' -Value {
+                return @{} | Add-Member -MemberType 'ScriptMethod' -Name 'Refresh' -Value {
+                    # Mock implementation
+                } -PassThru -Force
+            } -Force
+
+            $result = New-SqlDscDatabase -ServerObject $mockServerObject2022 -Name 'LedgerDatabase' -IsLedger $true -Force
+
+            $result | Should -Not -BeNullOrEmpty
+            $result.Name | Should -Be 'LedgerDatabase'
+            $result.IsLedger | Should -BeTrue
         }
 
         It 'Should throw error when database already exists' {
@@ -151,13 +170,23 @@ Describe 'New-SqlDscDatabase' -Tag 'Public' {
             $errorRecord.CategoryInfo.Category | Should -Be 'InvalidArgument'
             $errorRecord.CategoryInfo.TargetName | Should -Be 'InvalidCollation'
         }
+
+        It 'Should throw error when IsLedger is used on SQL Server version older than 2022' {
+            $errorRecord = { New-SqlDscDatabase -ServerObject $mockServerObject -Name 'TestDB' -IsLedger $true -Force } |
+                Should -Throw -ExpectedMessage '*IsLedger is not supported*' -PassThru
+
+            $errorRecord.Exception.Message | Should -BeLike '*IsLedger is not supported*'
+            $errorRecord.FullyQualifiedErrorId | Should -Be 'NSD0007,New-SqlDscDatabase'
+            $errorRecord.CategoryInfo.Category | Should -Be 'InvalidOperation'
+            $errorRecord.CategoryInfo.TargetName | Should -Be 'True'
+        }
     }
 
     Context 'Parameter validation' {
         It 'Should have the correct parameters in parameter set Database' -ForEach @(
             @{
                 ExpectedParameterSetName = 'Database'
-                ExpectedParameters = '-ServerObject <Server> -Name <string> [-Collation <string>] [-CatalogCollation <CatalogCollationType>] [-CompatibilityLevel <string>] [-RecoveryModel <string>] [-OwnerName <string>] [-FileGroup <DatabaseFileGroupSpec[]>] [-Force] [-Refresh] [-WhatIf] [-Confirm] [<CommonParameters>]'
+                ExpectedParameters = '-ServerObject <Server> -Name <string> [-Collation <string>] [-CatalogCollation <CatalogCollationType>] [-CompatibilityLevel <string>] [-RecoveryModel <string>] [-OwnerName <string>] [-IsLedger <bool>] [-FileGroup <DatabaseFileGroupSpec[]>] [-Force] [-Refresh] [-WhatIf] [-Confirm] [<CommonParameters>]'
             }
         ) {
             $result = (Get-Command -Name 'New-SqlDscDatabase').ParameterSets |
@@ -202,6 +231,12 @@ Describe 'New-SqlDscDatabase' -Tag 'Public' {
             $parameterInfo = (Get-Command -Name 'New-SqlDscDatabase').Parameters['DatabaseSnapshotBaseName']
             $snapshotSetAttribute = $parameterInfo.Attributes | Where-Object { $_.ParameterSetName -eq 'Snapshot' }
             $snapshotSetAttribute.Mandatory | Should -BeTrue
+        }
+
+        It 'Should have IsLedger as a parameter in Database parameter set' {
+            $parameterInfo = (Get-Command -Name 'New-SqlDscDatabase').Parameters['IsLedger']
+            $databaseSetAttribute = $parameterInfo.Attributes | Where-Object { $_.ParameterSetName -eq 'Database' }
+            $databaseSetAttribute | Should -Not -BeNullOrEmpty
         }
     }